change logic for check refresh token
